make hwcaps_t a pointer to uint64_t
  * aarch64/include/mach/aarch64/mach_aarch64.defs: (hwcaps_t) make type an
    array of uint64_t.
  * aarch64/include/mach/aarch64/mach_aarch64_types.h (hwcaps_t) make type a
    pointer to uint64_t

Acked-by: Sergey Bugaev <bugaevc@gmail.com>

Message-ID: <20250109224057.1652-1-dnietoc@gmail.com>

+++ aarch64/include/mach/aarch64/mach_aarch64_types.h
@@ -25,7 +25,7 @@
 
 #ifndef __ASSEMBLER__
 #include <stdint.h>
-typedef uint32_t *hwcaps_t;
+typedef uint64_t *hwcaps_t;
 #endif
 
 /* These definitions are meant to match those in
